Integrator

WinchControllers.Integrator — Type

mutable struct Integrator

Discrete integrator with external reset.

Fields

dt::Float64

i::Float64: Default: 1.0

output::Float64: Default: 0.0

last_output::Float64: Default: 0.0

WinchControllers.Integrator — Type

function Integrator(dt, i=1.0, x0=0.0)

Constructor for discrete integrator with external reset.

Parameters

dt: timestep [s]

i: integration constant

x0: initial and last output

Returns

a new struct of type Integrator

WinchControllers.reset — Function

reset(int::Integrator, x0=0.0)

Reset the integrator int to the value x0.

Parameters

int::Integrator: An integrator struct

x0: default value =0.0; initial and last output

Returns

nothing

WinchControllers.calc_output — Method

calc_output(int::Integrator, input)

Calculate and return the output without updating last_output.

Parameters

int::Integrator: An integrator struct

input: The input value

Returns

the output value

WinchControllers.on_timer — Method

on_timer(int::Integrator)

Update the field last_output. Must be called once per time-step.

Parameters

int::Integrator: An integrator struct

Returns

nothing

UnitDelay

WinchControllers.UnitDelay — Type

mutable struct UnitDelay

UnitDelay, delay the input signal by one time step.

Fields

last_output::Float64: Default: 0

last_input::Float64: Default: 0

WinchControllers.reset — Method

reset(ud::UnitDelay)

Reset the last_input and last_output of the struct ud to zero.

Parameters

ud::UnitDelay: A UnitDelay struct

Returns

nothing

WinchControllers.calc_output — Method

calc_output(ud::UnitDelay, input)

Calculate and return the output and update the last_input, but not the last_output.

Parameters

int::UnitDelay: A UnitDelay struct

input: The input value

Returns

the last output

WinchControllers.on_timer — Method

on_timer(ud::UnitDelay)

Update the field last_output. Must be called once per time-step.

Parameters

ud::UnitDelay: A UnitDelay struct

Returns

nothing

CalcVSetIn

WinchControllers.set_vset_pc — Method

set_vset_pc(cvi::CalcVSetIn, v_set_pc, force)

Parameters:

force: measured tether force [N]

v_set_pc: only used during manual operation or park-at-length. If it is nothing, v_set_in is calculated as function of the force.

Returns:

nothing

WinchControllers.calc_output — Method

calc_output(cvi::CalcVSetIn)

Parameters

cvi::CalcVSetIn: A struct of type CalcVSetIn

Returns

v_set_in: Either v_set, or a value, proportional to the square root of the force.