Input functions

WinchControllers.set_vset_pc — Function

set_vset_pc(cvi::CalcVSetIn, v_set_pc, force)

Parameters:

force: measured tether force [N]

vsetpc: only used during manual operation or park-at-length. If it is nothing, vsetin is calculated as function of the force.

WinchControllers.reset — Function

reset(int::Integrator, x0=0.0)

Reset the integrator int to the value x0.

Parameters

int::Integrator: An integrator struct

x0: default value =0.0; initial and last output

Output functions

WinchControllers.calc_output — Function

calc_output(cvi::CalcVSetIn)

Returns vsetin: Either v_set, or a value, proportional to the sqare root of the force.

WinchControllers.calc_vro — Function

Calculate the optimal reel-out speed for a given force.

WinchControllers.get_state — Function

Return the controller state as integer.

wcsLowerForceControl = 0 # input b selected wcsSpeedControl = 1 # input a selected wcsUpperForceControl = 2 # input c selected

Utility functions

WinchControllers.@limit — Macro

Limit the value of a variable.

Usage: @limit x 1 4 # limits the value to the range 1 .. 4, modifies x @limit x 10 # limits the value to the range -inf .. 10, modifies x

WinchControllers.saturate — Function

Calculate a saturated value, that stays within the given limits.

WinchControllers.merge_angles — Function

merge_angles(alpha, beta, factor_beta)

Calculate the weighted average of two angles. The weight of beta, factor_beta must be between 0 and 1.